scraper/requirements.txt
Crawl4AI
python-dotenv==1.0.1
pydantic==2.10.6
google-cloud-storage
openai
beautifulsoup4
groq